Moves files over from the project-latex repo
Adds package.json so that we explicitly declcare the dependencies of the
project.

Adds telemetryKeys.json into the project. It's not ideal that this means
we'll have duplicate copies in this repo and the balloon repo, but it
feels worth the issue so that we can deploy this project to Heroku.

Also renames test spec so it matches the format that jasmine-node expects.